Some minor changes
